Authorize thumbnail requests
Closes #22. This fixes thumbnail loading by passing payload/sig to the
CDN. Authorization can now be passed via headers or query parameters.
Authorization is also passed to the storyboard by replacing the file
image address with the authorized payload/sig on the fly. This was hacky
but vidstack doesn't provide a way to edit the storyboard request and
the only other way is using service workers to modify the request